pyproject.toml
[tool.poetry]
name = "flask-project"
version = "0.1.0"
description = "Flask werkzeug"
authors = ["kpomak <romka_ne@mail.ru>"]
license = "MIT"
readme = "README.md"
packages = [{include = "flask_project"}]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^2.2.2"
flask-sqlalchemy = "^3.0.2"
flask-login = "^0.6.2"
flask-migrate = "^4.0.0"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
black = "^22.12.0"
isort = "^5.11.4"

[tool.poetry.group.dev.dependencies]
Flask = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
